Add files via upload
SPORTBET, a professional sports betting Solution that comes with ZEND PHP laravel. It’s developed for those people who want to start their Bettings website. Sports betting makes up about 30 to 40 percent of the global gambling market, which also includes lotteries, casinos, poker and other gaming. It's estimated that over half a trillion (over €500 million) is wagered in bets around the world annually. SPORTBET, a complete solution for Betting business. here, user join directly or by reference and able to bet on any categories, live sports or game. admin able to control all users, website and manage game, live bettings and able to select winner.

The documentation will help to understand the total system and we have tried to summarize the total process of the site here. This script comes with a very simple and easy-to-use admin panel, user panel which leads to use without any coding knowledge. Additionally, this script comes with many Automated online payment gateways to make payment easier all over the world.

If needed, we will update this script continuously and come with new features and security patches. We will try to give the best user experience to all of our clients.

Thank you for purchasing SPORTBET. If you have any questions or queries that are not answered in this document, please feel free to contact us via email. We will try to respond to you as soon as possible. Thank you so much.